pyproject.toml
[project]
name = "buylog"
version = "0.1.9"
description = "A purchasing support tool with CLI and TUI interfaces for vendor quote management."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Shakeeb Alireza", email = "shakfu@users.noreply.github.com" }
]
maintainers = [
    { name = "Shakeeb Alireza", email = "shakfu@users.noreply.github.com" }
]
requires-python = ">=3.13"
keywords = [
    "purchasing",
    "quotes",
    "vendor-management",
    "price-tracking",
    "cli",
    "tui",
    "textual",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "sqlalchemy>=2.0.40",
    "tabulate>=0.9.0",
    "textual>=7.3.0",
    "textual-dev>=1.8.0",
    "pyperclip>=1.9.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0.3",
    "openpyxl>=3.1.5",
]

[project.urls]
Homepage = "https://github.com/shakfu/buylog"
Repository = "https://github.com/shakfu/buylog"
Issues = "https://github.com/shakfu/buylog/issues"
Changelog = "https://github.com/shakfu/buylog/blob/main/CHANGELOG.md"

[project.scripts]
buylog = "buylog:main"

[build-system]
requires = ["uv_build>=0.7.3,<0.8"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "eralchemy>=1.5.0",
    "factory-boy>=3.3.3",
    "litecli>=1.15.0",
    "mypy>=1.19.1",
    "py-pglite[sqlalchemy]>=0.5.3",
    "pytest>=8.3.5",
    "ruff>=0.14.13",
    "twine>=6.2.0",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
filterwarnings = [
    # SQLAlchemy uses deprecated utcnow() internally - upstream issue
    "ignore:datetime.datetime.utcnow:DeprecationWarning",
]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
warn_unused_ignores = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "pyperclip",
    "requests",
    "tabulate",
    "bs4",
]
ignore_missing_imports = true
